+++ src/DatasetBase/Support/AbstractCollection.php
@@ -255,4 +255,26 @@ abstract class AbstractCollection implements CollectionInterface
 
         return $that;
     }
+
+    /**
+     * Alternative implementation for @see contains to check contains by more detailed object comparision.
+     * This is useful, when the collection contains items that can be equal even if they are not identical.
+     *
+     * @param T         $value
+     * @param Closure(T $a,    T $b): bool $equalsCondition
+     */
+    final protected function containsByEqualsCheck($value, \Closure $equalsCondition): bool
+    {
+        if (!$this->isValidItem($value)) {
+            return false;
+        }
+
+        foreach ($this->items as $item) {
+            if ($equalsCondition($item, $value)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
 }