+++ src/PortalBase/Mapping/MappingComponentCollection.php
@@ -15,6 +15,16 @@ use Heptacom\HeptaConnect\Portal\Base\StorageKey\PortalNodeKeyCollection;
  */
 class MappingComponentCollection extends AbstractObjectCollection
 {
+    public function contains($value): bool
+    {
+        return $this->containsByEqualsCheck(
+            $value,
+            static fn (MappingComponentStructContract $a, MappingComponentStructContract $b): bool => $a->getPortalNodeKey()->equals($b->getPortalNodeKey())
+                && $a->getEntityType() === $b->getEntityType()
+                && $a->getExternalId() === $b->getExternalId()
+        );
+    }
+
     /**
      * @psalm-return class-string<DatasetEntityContract>[]
      *