+++ src/DatasetBase/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add method `\Heptacom\HeptaConnect\Dataset\Base\Contract\CollectionInterface::reverse` and implementation in `\Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection::reverse` to get a reversed collection
 - Add method `\Heptacom\HeptaConnect\Dataset\Base\Contract\CollectionInterface::isEmpty` and implementation in `\Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection::isEmpty` to check whether the collection is empty without counting
 - Add method `\Heptacom\HeptaConnect\Dataset\Base\Contract\CollectionInterface::contains` and implementation in `\Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection::contains` to check whether the given item is in the collection
+- Add method `\Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection::containsByEqualsCheck` for any extending class to build alternative contains implementations based upon comparison
 
 ### Changed